ScoreHandler.cs work. Changed ChartNote.Position to float.

+++ Assets/RhythmGameAssets/Scripts/ChartNote.cs
@@ -11,11 +11,11 @@ namespace RhythmGameAssets.Scripts
         public int ID { get; set; }
         // public static int ID { get; set; }
         public NoteDirection Direction { get; set; }
-        public int Position { get; set; } // the time the note appears in ticks
-        public int Length { get; set; } // length in ticks
+        public double Position { get; set; } // the time the note appears in ticks
+        public double Length { get; set; } // length in ticks
         public ChartNote Next { get; set; }
 
-        public ChartNote(int id, NoteDirection direction, int position, int length)
+        public ChartNote(int id, NoteDirection direction, double position, double length)
         {
             ID = id;
             Direction = direction;
